requirements.txt
asgiref==3.5.0
azure-cognitiveservices-vision-customvision==3.1.0
azure-cognitiveservices-vision-face==0.6.0
azure-common==1.1.27
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.2.11
django-crispy-forms==1.14.0
django-environ==0.8.1
idna==3.3
isodate==0.6.1
msrest==0.6.21
numpy==1.21.5
oauthlib==3.1.1
opencv-python==4.5.5.62
Pillow==9.0.0
pyaadhaar==1.1.2
pylibjpeg==1.4.0
pylibjpeg-openjpeg==1.2.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
pyzbar==0.1.8
qrtools==0.0.2
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.8
xmltodict==0.12.0
